agent/pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "AI Task Assistant Agent with multi-provider LLM support"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "openai-agents>=0.9.1",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
litellm = [
    "litellm>=1.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]
